requirements.txt
streamlit>=1.28.0
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
scipy>=1.11.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
flake8>=6.0.0
